requirements.txt
sentence-transformers==2.7.0
chromadb==0.4.24
pypdf==4.3.1
tqdm==4.66.4
numpy>=1.24
groq>=0.9.0,<1.0.0
duckduckgo-search>=6.2.6
python-dotenv>=1.0.1
psycopg[binary]>=3.1
